test: move OpenWrt probe regression to Swift Testing
