+++ cmuxTests/WorkspaceRemoteConnectionTests.swift
@@ -1762,87 +1762,6 @@ final class WorkspaceRemoteConnectionTests: XCTestCase {
         XCTAssertEqual(remotePath, "/tmp/cmux-drop-12345678-1234-1234-1234-1234567890ab.png")
     }
 
-    func testRemotePlatformProbeScriptWorksWhenTrLacksCharacterClasses() throws {
-        let fileManager = FileManager.default
-        let root = fileManager.temporaryDirectory.appendingPathComponent(
-            "cmux-remote-platform-probe-\(UUID().uuidString)",
-            isDirectory: true
-        )
-        let home = root.appendingPathComponent("home", isDirectory: true)
-        let bin = root.appendingPathComponent("bin", isDirectory: true)
-        let daemonURL = home
-            .appendingPathComponent(".cmux/bin/cmuxd-remote/test-version/linux-amd64", isDirectory: true)
-            .appendingPathComponent("cmuxd-remote", isDirectory: false)
-        try fileManager.createDirectory(at: daemonURL.deletingLastPathComponent(), withIntermediateDirectories: true)
-        try fileManager.createDirectory(at: bin, withIntermediateDirectories: true)
-        defer { try? fileManager.removeItem(at: root) }
-
-        try writeExecutableShellFile(
-            at: daemonURL,
-            body: """
-            #!/bin/sh
-            exit 0
-            """
-        )
-        try writeExecutableShellFile(
-            at: bin.appendingPathComponent("uname"),
-            body: """
-            #!/bin/sh
-            case "${1:-}" in
-              -s) printf '%s\\n' Linux ;;
-              -m) printf '%s\\n' x86_64 ;;
-              *) exit 1 ;;
-            esac
-            """
-        )
-        try writeExecutableShellFile(
-            at: bin.appendingPathComponent("tr"),
-            body: """
-            #!/bin/sh
-            if [ "$#" -eq 2 ] && [ "$1" = '[:upper:]' ] && [ "$2" = '[:lower:]' ]; then
-              awk -v from="$1" -v to="$2" '
-                BEGIN {
-                  for (i = 1; i <= length(from); i++) {
-                    map[substr(from, i, 1)] = substr(to, i, 1)
-                  }
-                }
-                {
-                  if (NR > 1) {
-                    printf "\\n"
-                  }
-                  output = ""
-                  for (i = 1; i <= length($0); i++) {
-                    ch = substr($0, i, 1)
-                    output = output ((ch in map) ? map[ch] : ch)
-                  }
-                  printf "%s", output
-                }
-              '
-              exit 0
-            fi
-            exec /usr/bin/tr "$@"
-            """
-        )
-
-        let result = runProcess(
-            executablePath: "/usr/bin/env",
-            arguments: [
-                "HOME=\(home.path)",
-                "PATH=\(bin.path):/usr/bin:/bin",
-                "/bin/sh",
-                "-c",
-                WorkspaceRemoteSessionController.remotePlatformProbeScript(version: "test-version"),
-            ],
-            timeout: 5
-        )
-
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stdout + result.stderr)
-        XCTAssertTrue(result.stdout.contains("__CMUX_REMOTE_OS__=Linux"), result.stdout)
-        XCTAssertTrue(result.stdout.contains("__CMUX_REMOTE_ARCH__=x86_64"), result.stdout)
-        XCTAssertTrue(result.stdout.contains("__CMUX_REMOTE_EXISTS__=yes"), result.stdout)
-    }
-
     @MainActor
     func testDaemonBootstrapUploadUsesAbsoluteHomePathForScpDestination() throws {
         let fileManager = FileManager.default
